Revert "Upg: mutate() provided by useSWRDefaults now automatically mutate all…"
This reverts commit 95459c72d7a27371158a5dd3080bcc070b28f6c3.
