+++ front/components/vaults/VaultResourcesList.tsx
@@ -12,6 +12,7 @@ import {
   usePaginationFromUrl,
 } from "@dust-tt/sparkle";
 import type {
+  APIError,
   ConnectorProvider,
   DataSourceViewCategory,
   DataSourceViewType,
@@ -24,23 +25,21 @@ import { isWebsiteOrFolderCategory } from "@dust-tt/types";
 import type { CellContext, ColumnDef } from "@tanstack/react-table";
 import { useRouter } from "next/router";
 import type { ComponentType } from "react";
-import { useMemo } from "react";
+import { useContext, useMemo } from "react";
 import { useRef } from "react";
 import { useState } from "react";
 
 import { ConnectorPermissionsModal } from "@app/components/ConnectorPermissionsModal";
 import ConnectorSyncingChip from "@app/components/data_source/DataSourceSyncChip";
 import { DeleteStaticDataSourceDialog } from "@app/components/data_source/DeleteStaticDataSourceDialog";
+import { SendNotificationsContext } from "@app/components/sparkle/Notification";
 import { AddConnectionMenu } from "@app/components/vaults/AddConnectionMenu";
 import { EditVaultManagedDataSourcesViews } from "@app/components/vaults/EditVaultManagedDatasourcesViews";
 import { EditVaultStaticDatasourcesViews } from "@app/components/vaults/EditVaultStaticDatasourcesViews";
 import { getConnectorProviderLogoWithFallback } from "@app/lib/connector_providers";
 import { getDataSourceNameFromView } from "@app/lib/data_sources";
 import { useDataSources } from "@app/lib/swr/data_sources";
-import {
-  useDeleteFolderOrWebsite,
-  useVaultDataSourceViews,
-} from "@app/lib/swr/vaults";
+import { useVaultDataSourceViews } from "@app/lib/swr/vaults";
 import { classNames } from "@app/lib/utils";
 
 const REDIRECT_TO_EDIT_PERMISSIONS = [
@@ -210,6 +209,7 @@ export const VaultResourcesList = ({
   const [isNewConnectorLoading, setIsNewConnectorLoading] = useState(false);
   const { dataSources, isDataSourcesLoading } = useDataSources(owner);
   const router = useRouter();
+  const sendNotification = useContext(SendNotificationsContext);
 
   const searchBarRef = useRef<HTMLInputElement>(null);
 
@@ -223,12 +223,6 @@ export const VaultResourcesList = ({
 
   const { pagination, setPagination } = usePaginationFromUrl({
     urlPrefix: "table",
-  });
-
-  const doDelete = useDeleteFolderOrWebsite({
-    owner,
-    vaultId: vault.sId,
-    category,
   });
 
   // DataSources Views of the current vault.
@@ -311,14 +305,32 @@ export const VaultResourcesList = ({
   }
 
   const onDeleteFolderOrWebsite = async () => {
-    if (selectedDataSourceView?.dataSource) {
-      const res = await doDelete(selectedDataSourceView.dataSource);
-      if (res) {
-        await router.push(
-          `/w/${owner.sId}/vaults/${vault.sId}/categories/${selectedDataSourceView.category}`
-        );
-      }
+    if (!selectedDataSourceView) {
+      return;
     }
+    const res = await fetch(
+      `/api/w/${owner.sId}/vaults/${vault.sId}/data_sources/${selectedDataSourceView.dataSource.name}`,
+      { method: "DELETE" }
+    );
+
+    if (res.ok) {
+      await router.push(
+        `/w/${owner.sId}/vaults/${vault.sId}/categories/${selectedDataSourceView.category}`
+      );
+      sendNotification({
+        type: "success",
+        title: `Successfully deleted ${selectedDataSourceView.category}`,
+        description: `${getDataSourceNameFromView(selectedDataSourceView)} was successfully deleted.`,
+      });
+    } else {
+      const err: { error: APIError } = await res.json();
+      sendNotification({
+        type: "error",
+        title: `Error deleting ${selectedDataSourceView.category}`,
+        description: `Error: ${err.error.message}`,
+      });
+    }
+    return res.ok;
   };
 
   return (
