+++ front/components/vaults/VaultSideBarMenu.tsx
@@ -470,6 +470,7 @@ const VaultDataSourceViewSubMenu = ({
       workspaceId: owner.sId,
       vaultId: vault.sId,
       category,
+      disabled: !isExpanded,
     });
 
   return (
@@ -482,11 +483,6 @@ const VaultDataSourceViewSubMenu = ({
       onChevronClick={() => setIsExpanded(!isExpanded)}
       visual={categoryDetails.icon}
       areActionsFading={false}
-      type={
-        isVaultDataSourceViewsLoading || vaultDataSourceViews.length > 0
-          ? "node"
-          : "leaf"
-      }
     >
       {isExpanded && (
         <Tree isLoading={isVaultDataSourceViewsLoading}>
