+++ front/lib/swr/swr.ts
@@ -1,5 +1,4 @@
 import type { PaginationState } from "@tanstack/react-table";
-import { useCallback } from "react";
 import type { Fetcher, Key, SWRConfiguration } from "swr";
 import useSWR, { useSWRConfig } from "swr";
 
@@ -14,69 +13,19 @@ export function useSWRWithDefaults<TKey extends Key, TData>(
   fetcher: Fetcher<TData, TKey> | null,
   config?: SWRConfiguration & { disabled?: boolean }
 ) {
-  const { mutate: globalMutate, cache } = useSWRConfig();
+  const { mutate: globalMutate } = useSWRConfig();
 
   const mergedConfig = { ...DEFAULT_SWR_CONFIG, ...config };
   const disabled = !!mergedConfig.disabled;
 
   const result = useSWR(disabled ? null : key, fetcher, mergedConfig);
 
-  // If the key looks like an url, we need to remove the query params
-  // to make sure we don't cache different pages together
-  // Naive way to detect url by checking for '/'
-  const tryMakeUrlWithoutParams = useCallback((key: TKey) => {
-    if (typeof key === "string" && key.includes("/")) {
-      try {
-        const urlFromKey = new URL(
-          key,
-          key.indexOf("://") == -1 ? "https://example.org/" : undefined // We need to provide a base url to make sure the URL is parsed correctly
-        );
-        return urlFromKey.origin + urlFromKey.pathname;
-      } catch {
-        return null;
-      }
-    } else {
-      return null;
-    }
-  }, []);
-
-  const mutateKeysWithSameUrl = useCallback(
-    (key: TKey) => {
-      const keyAsUrl = tryMakeUrlWithoutParams(key);
-      if (keyAsUrl) {
-        // Cycle through all the keys in the cache that start with the same url
-        // and mutate them too
-        for (const k of cache.keys()) {
-          const kAsUrl = tryMakeUrlWithoutParams(k as TKey);
-          if (kAsUrl === keyAsUrl) {
-            void globalMutate<TData>(k);
-          }
-        }
-      }
-    },
-    [tryMakeUrlWithoutParams, cache, globalMutate]
-  );
-
   if (disabled) {
     // When disabled, as the key is null, the mutate function is not working
     // so we need to provide a custom mutate function that will work
-    return {
-      ...result,
-      mutate: () => {
-        mutateKeysWithSameUrl(key);
-        return globalMutate(key);
-      },
-    };
+    return { ...result, mutate: () => globalMutate<TData>(key) };
   } else {
-    const myMutate: typeof result.mutate = (data, opts) => {
-      mutateKeysWithSameUrl(key);
-      return result.mutate(data, opts);
-    };
-
-    return {
-      ...result,
-      mutate: myMutate,
-    };
+    return result;
   }
 }
